- New class OEntityProxy impersonates uninstantiated entities during   registration, and also holds meta information necessary to instantiate   correctly. - New protocol OEntityFacade in combination with method forwarding   facilitates access to defined properties across entity and proxy   incarnations of entities. - Reclassified OAddressFacade as OOrigoProxy, subclassing OEntityProxy. - Streamlined state transition between view controllers, now explicitly   passing on entity (or proxy) reference as target for next view   controller. - New OTableViewControllerInstance method -defaultTarget facilitates   setting of default target when no target is transferred from parent view   controller (e.g., when there is no parent view controller, or for non-   entity view controllers). - Now longer permitting picking of cross-generational members for origo   inclusion. - Now facilitating pre-entity insertion editing of addresses returned from   address book during member registration. - Cleaned up OOrigo and OMember category methods, now primarily returning   sets of counterparts instead of memberships. - Fixed bug whereby entities were replicated twice concurrently when   registering new origos.

+++ OrigoApp/OCrypto.m
@@ -3,7 +3,7 @@
 //  OrigoApp
 //
 //  Created by Anders Blehr on 17.10.12.
-//  Copyright (c) 2012 Rhelba Creations. All rights reserved.
+//  Copyright (c) 2012 Rhelba Source. All rights reserved.
 //
 
 #import "OCrypto.h"
@@ -33,7 +33,7 @@ static NSInteger const kActivationCodeLength = 6;
     char seasonedBytes[hashLength + 1];
     seasonedBytes[hashLength] = (char)0;
     
-    for (int i = 0; i < hashLength; i++) {
+    for (NSInteger i = 0; i < hashLength; i++) {
         char char1 = selfCString[i];
         char char2 = seasoningCString[hashLength - (i + 1)];
         
@@ -146,7 +146,7 @@ static NSInteger const kActivationCodeLength = 6;
     
     NSMutableString *hash = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
     
-    for (int i = 0; i < CC_SHA1_DIGEST_LENGTH; i++) {
+    for (NSInteger i = 0; i < CC_SHA1_DIGEST_LENGTH; i++) {
         [hash appendFormat:@"%02x", SHA1Digest[i]];
     }
